Ability to automatically expose relations for a resource
As per specification a resource may choose to expose relations within a
links node (this happens currently when including a relation)

This PR adds an attribute to the model which enables a resource to
identify relations it wishes to expose via the links permanently (when
no include is specified).

This identifies the resource within the links node but DOES NOT include
the relation.

http://jsonapi.org/format/#document-structure-resource-relationships